feat: 프로젝트 목록 조회 및 진행률/상태 스타일링 기능 구현
- 프로젝트 목록 조회 시 로그인 사용자 기준 필터 적용 (PROJECT_USER_GROUP 테이블 조인)
- PJT 조회 쿼리 페이징 처리 및 검색 조건 적용
- 프로젝트 등록 후 생성된 PJT_ID 반환 처리 (selectKey 활용)
- WebSquare 화면에서 사용자 정보 전역 변수 처리 및 onpageload 초기 설정
- 프로젝트 목록 조회 후 진행률 계산(setProgressRate) 및 컬럼 스타일 적용(setProgressStyle)
- 상태(pjtStatus)에 따라 상태 색상 원 아이콘 표시(setStatusStyle)
- WebSquare 스크립트에 W2Script 표준 주석 형식 적용
