+++ src/main/java/com/demo/proworks/pjt/vo/PjtListVo.java
@@ -1,4 +1,4 @@
-package com.demo.proworks.stg.vo;
+package com.demo.proworks.pjt.vo;
 
 import com.inswave.elfw.annotation.ElDto;
 import com.inswave.elfw.annotation.ElDtoField;
@@ -9,14 +9,14 @@ import com.fasterxml.jackson.annotation.JsonFilter;
 public class PjtListVo extends com.demo.proworks.cmmn.ProworksCommVO {
     private static final long serialVersionUID = 1L;
 
-    @ElDtoField(logicalName = "프로젝트 정보List", physicalName = "pjtVoList", type = "com.demo.proworks.stg.PjtVo", typeKind = "List", fldYn = "", length = 0, dotLen = 0, baseValue = "", desc = "")
-    private java.util.List<com.demo.proworks.stg.vo.PjtVo> pjtVoList;
+    @ElDtoField(logicalName = "프로젝트 정보List", physicalName = "pjtVoList", type = "com.demo.proworks.pjt.PjtVo", typeKind = "List", fldYn = "", length = 0, dotLen = 0, baseValue = "", desc = "")
+    private java.util.List<com.demo.proworks.pjt.vo.PjtVo> pjtVoList;
 
-    public java.util.List<com.demo.proworks.stg.vo.PjtVo> getPjtVoList(){
+    public java.util.List<com.demo.proworks.pjt.vo.PjtVo> getPjtVoList(){
         return pjtVoList;
     }
 
-    public void setPjtVoList(java.util.List<com.demo.proworks.stg.vo.PjtVo> pjtVoList){
+    public void setPjtVoList(java.util.List<com.demo.proworks.pjt.vo.PjtVo> pjtVoList){
         this.pjtVoList = pjtVoList;
     }
