--- src/main/resources/inswave/sqlmap/com/demo/proworks/stg/pjt_SQL_mariadb_MyBatis.xml
@@ -1,127 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper      
-    PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"      
-    "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<!--
-	@subject     : 프로젝트 정보 관련 처리를 담당하는 Sql Mapper
-	@description : 프로젝트 정보 관련 처리를 담당하는 Sql Mapper
-	@author      : 김성민
-	@since       : 2025/07/05
-	@modification
-	===========================================================
-	DATE              AUTHOR             DESC
-	===========================================================
-	2025/07/05              김성민             최초 생성
- -->
-<mapper namespace="com.demo.proworks.stg">
-    <insert id="insertPjt" parameterType="com.demo.proworks.stg.vo.PjtVo">
-    	<!-- 프로젝트 정보를 등록한다. -->	
-    		
-        INSERT INTO PJT   /* QueryID : com.demo.proworks.stg.insertPjt */
-        ( 
-            PJT_ID
-            <if test="pjtName  != null"> , PJT_NAME   </if>
-            <if test="pjtAsi   != null"> , PJT_ASI    </if>
-            <if test="pjtStatus!= null"> , PJT_STATUS </if>
-            <if test="pjtDetail!= null"> , PJT_DETAIL </if>
-            <if test="pjtSt    != null and pjtSt != '' "> , PJT_ST     </if>
-            <if test="pjtEt    != null and pjtEt != '' "> , PJT_ET     </if>
-        )
-        VALUES  
-        ( 
-            #{pjtId}
-            <if test="pjtName  != null"> , #{pjtName}   </if>
-            <if test="pjtAsi   != null"> , #{pjtAsi}    </if>
-            <if test="pjtStatus!= null"> , #{pjtStatus} </if>
-            <if test="pjtDetail!= null"> , #{pjtDetail} </if>
-            <if test="pjtSt    != null and pjtSt != '' "> , #{pjtSt}     </if>
-            <if test="pjtEt    != null and pjtEt != '' "> , #{pjtEt}     </if> 
-        )          
-    </insert>	
-	
-    <update id="updatePjt" parameterType="com.demo.proworks.stg.vo.PjtVo">
-    	<!-- 프로젝트 정보를 갱신한다. -->
-    	
-        UPDATE PJT    /* QueryID : com.demo.proworks.stg.updatePjt */      
-      	  SET  
-                PJT_ID     = #{pjtId}            	                       
-            <if test="pjtName  != null"> , PJT_NAME   = #{pjtName}   </if>
-            <if test="pjtAsi   != null"> , PJT_ASI    = #{pjtAsi}    </if>
-            <if test="pjtStatus!= null"> , PJT_STATUS = #{pjtStatus} </if>
-            <if test="pjtDetail!= null"> , PJT_DETAIL = #{pjtDetail} </if>
-            <if test="createdAt!= null"> , CREATED_AT = #{createdAt} </if>
-            <if test="updatedAt!= null"> , UPDATED_AT = #{updatedAt} </if>
-            <if test="pjtSt    != null"> , PJT_ST     = #{pjtSt}     </if>
-            <if test="pjtEt    != null"> , PJT_ET     = #{pjtEt}     </if>                 
-        WHERE   
-                PJT_ID     = #{pjtId}    
-    </update>
-	
-    <delete id="deletePjt" parameterType="com.demo.proworks.stg.vo.PjtVo">
-    	<!-- 프로젝트 정보를 삭제한다. -->
-    	
-        DELETE FROM PJT   /* QueryID : com.demo.proworks.stg.deletePjt */ 
-        WHERE   
-                PJT_ID     = #{pjtId}            
-    </delete>
-	
-    <select id="selectPjt" parameterType="com.demo.proworks.stg.vo.PjtVo" resultType="com.demo.proworks.stg.vo.PjtVo">
-    	<!-- 프로젝트 정보를 상세 조회한다. -->	
-        SELECT     /* QueryID : com.demo.proworks.stg.selectPjt */
-            PJT_ID as pjtId, PJT_NAME as pjtName, PJT_ASI as pjtAsi, PJT_STATUS as pjtStatus, PJT_DETAIL as pjtDetail, CREATED_AT as createdAt, UPDATED_AT as updatedAt, PJT_ST as pjtSt, PJT_ET as pjtEt
-        FROM PJT 
-        WHERE 
-                PJT_ID     = #{pjtId}    			
-    </select>
-	
-    <select id="selectListPjt" parameterType="com.demo.proworks.stg.vo.PjtVo" resultType="com.demo.proworks.stg.vo.PjtVo">
-    	<!-- 프로젝트 정보 목록을 조회한다. -->
-           SELECT    /* QueryID : com.demo.proworks.stg.selectListPjt */
-               PJT_ID as pjtId, PJT_NAME as pjtName, PJT_ASI as pjtAsi, PJT_STATUS as pjtStatus, PJT_DETAIL as pjtDetail, CREATED_AT as createdAt, UPDATED_AT as updatedAt, PJT_ST as pjtSt, PJT_ET as pjtEt		       
-           FROM (
-               SELECT  S.*, CEIL( (@ROWNUM:=@ROWNUM+1) / #{pageSize} ) AS PAGE
-               FROM (				
-                   SELECT 
-                       *  		                  
-                   FROM PJT
-                   WHERE  1=1    		
-                        <if test="scPjtId != null and scPjtId != ''">
-                             AND PJT_ID like '%${scPjtId}%'
-                        </if>
-                        <if test="scPjtName != null and scPjtName != ''">
-                             AND PJT_NAME like '%${scPjtName}%'
-                        </if>
-                        <if test="scPjtAsi != null and scPjtAsi != ''">
-                             AND PJT_ASI like '%${scPjtAsi}%'
-                        </if>
-                        <if test="scPjtStatus != null and scPjtStatus != ''">
-                             AND PJT_STATUS like '%${scPjtStatus}%'
-                        </if>	     																														
-               ) S, (SELECT @ROWNUM := 0) TMP
-           ) A
-            WHERE PAGE = #{pageIndex}
-    </select>
-	
-    <select id="selectListCountPjt" parameterType="com.demo.proworks.stg.vo.PjtVo" resultType="long">
-    <!-- 프로젝트 정보 목록의 카운트를 조회한다. -->
-    
-         SELECT    /* QueryID : com.demo.proworks.stg.selectListCountPjt */
-             COUNT(*) totcnt  		                  
-         FROM PJT
-         WHERE  1=1    		
-            <if test="scPjtId != null and scPjtId != ''">
-                 AND PJT_ID like '%${scPjtId}%'
-            </if>
-            <if test="scPjtName != null and scPjtName != ''">
-                 AND PJT_NAME like '%${scPjtName}%'
-            </if>
-            <if test="scPjtAsi != null and scPjtAsi != ''">
-                 AND PJT_ASI like '%${scPjtAsi}%'
-            </if>
-            <if test="scPjtStatus != null and scPjtStatus != ''">
-                 AND PJT_STATUS like '%${scPjtStatus}%'
-            </if>
-    </select>
-
-</mapper>
